Move DefaultKubeAPIServerSNI function to kubeapiserverexposure.go
