+++ pkg/operation/botanist/kubeapiserverexposure.go
@@ -21,6 +21,7 @@ import (
 	"github.com/gardener/gardener/pkg/operation/botanist/component"
 	"github.com/gardener/gardener/pkg/operation/botanist/component/kubeapiserverexposure"
 	"github.com/gardener/gardener/pkg/utils"
+	gutil "github.com/gardener/gardener/pkg/utils/gardener"
 
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -57,9 +58,36 @@ func (b *Botanist) DeployKubeAPIService(ctx context.Context, sniPhase component.
 	return b.newKubeAPIServiceServiceComponent(sniPhase).Deploy(ctx)
 }
 
+// DefaultKubeAPIServerSNI returns a deployer for kube-apiserver SNI.
+func (b *Botanist) DefaultKubeAPIServerSNI() component.DeployWaiter {
+	return component.OpDestroy(kubeapiserverexposure.NewSNI(
+		b.K8sSeedClient.Client(),
+		b.K8sSeedClient.Applier(),
+		b.Shoot.SeedNamespace,
+		&kubeapiserverexposure.SNIValues{
+			IstioIngressGateway:      b.getIngressGatewayConfig(),
+			APIServerInternalDNSName: b.outOfClusterAPIServerFQDN(),
+		},
+	))
+}
+
 func (b *Botanist) getKubeAPIServerServiceAnnotations(sniPhase component.Phase) map[string]string {
 	if b.ExposureClassHandler != nil && sniPhase != component.PhaseEnabled {
 		return utils.MergeStringMaps(b.Seed.LoadBalancerServiceAnnotations, b.ExposureClassHandler.LoadBalancerService.Annotations)
 	}
 	return b.Seed.LoadBalancerServiceAnnotations
+}
+
+func (b *Botanist) getIngressGatewayConfig() kubeapiserverexposure.IstioIngressGateway {
+	ingressGatewayConfig := kubeapiserverexposure.IstioIngressGateway{
+		Namespace: *b.Config.SNI.Ingress.Namespace,
+		Labels:    b.Config.SNI.Ingress.Labels,
+	}
+
+	if b.ExposureClassHandler != nil {
+		ingressGatewayConfig.Namespace = *b.ExposureClassHandler.SNI.Ingress.Namespace
+		ingressGatewayConfig.Labels = gutil.GetMandatoryExposureClassHandlerSNILabels(b.ExposureClassHandler.SNI.Ingress.Labels, b.ExposureClassHandler.Name)
+	}
+
+	return ingressGatewayConfig
 }
